未添加图片

+++ app/src/main/java/com/connor/picofull/models/HomeData.kt
@@ -4,7 +4,7 @@ data class HomeData(
     var waveId: Int? = null,
     var energy: Int = 0,
     var rate: Int = 0,
-    var pulse: Int = 0,
+    var pulse: Long = 0,
     var spot: Int = 0,
     var switch: Boolean = false,
     var readLight: Int = 0
